Reworked CSSProperties formatter (getClass function) into a common function and made minor changes to readback and led so they both used it
