+++ src/ui/widgets/LED/led.module.css
@@ -1,4 +1,4 @@
-.led {
+.Default {
   margin: 20px auto;
   width: 12px;
   height: 12px;
